we now got charts!
I cleaned up a few more bugs with the analyzer trying to get OHLC data
from weekend and holiday days. If this happens I just reset the
requested day back to the most recent trading day. I included the
highcharts and highstock frameworks for charting and stock data
visualization. An example is now showing up in the reports. Actual data
that reflects that stock will start popping up in the next update.